+++ lib/game/src/lib.rs
@@ -1,4 +1,7 @@
+pub mod animation;
+pub mod entity;
 pub mod event;
 pub mod job_class;
 pub mod movement;
 pub mod path;
+pub mod sprite_path;